Initial loader work.
The loader supports files in json, yaml, or toml format.
The example under examples/basicdemo includes files of each format
for testing purposes.

For the moment, loading works as such:
- In the top-level directory of a rules directory is a ruleset.*
  file in one of the supported formats. This is parsed into the
  camp.engine.base.Ruleset model class.
- The Ruleset has a `feature_model_def` field, which should contain
  the dotted name of a Python module and an attribute with it. The
  attribute should contain a union of Pydantic BaseModel subclasses.
- The loader will find all supported filetypes in subfolders and
  parse them into dictionaries. Files named __defaults__.* are special.
  They will be loaded first, and will be used for default field
  values for all files in the same directory or subdirectories.
- All features must have an `id` field populated with a string. By
  default, this will be the stem of the filename, though it can be
  specified manually if it needs to differ. This is more important
  when the file is a YAML stream document containing multiple features.
- The dictionaries will be parsed using the Pydantic model definition(s)
  given.
- All features that parse without error will be stored in the `features`
  attribute of the ruleset, which is a dictionary mapping `id` to the
  feature.
- Any feature that does not parse will be placed in the `bad_defs`
  attribute of the ruleset, which is a list of `BadDefinition` objects.
  A ruleset with non-empty `bad_defs` can't be used.

TODO: Allow markdown files to be parsed. For example, in basicdemo, there
are files named things like "fighter.description.md". Since the main
definition for `fighter` is in a JSON file, including long descriptive
text isn't convenient, so the idea that if a file is named something
like "A.B.md", the dictionary for A will have its key B populated
with the contents before parsing. Markdown files are just one possibility.

Later, we'll have to get around to actually doing something with these
feature definitions.