Add the Aggregator.
The aggregator collects statistics about everything on a character sheet.
The character model and each feature type can add their statistics to it
in whatever way they need to.

Most requirements take the form of an identifier (indicating a feature,
attribute, currency, or some aggregate concept) and optionally a number
that serves as a minimum. There are some variations on this, primarily
when dealing with classes:

* Do you have at least 5 levels of Mage?
* Do you have at least 5 levels of any casting class?
* Do you have any levels of any non-casting class?
* Do you have any single class with at least 13 levels?

The geas5 ClassDef feature model defineas an `aggregate` method that
fills the aggregator with appropriate data. If no custome aggregate
method is defined, a default is provided that handles the common case
of a basic feature that may have ranks and options. For example, if you
have "Lore [Undead]" and "Lore [Arcane]", you'll get:

* lore = 2
* lore#Undead = 1
* lore#Arcane = 1
* :max_lore = 1

Note that the "basicdemo" system is removed. For now I'm focusing on geas5
as the canonical system and don't want to have to think about overly general
hypothetical cases. We'll worry about that later.