Add `decrease` for skills. Sort of.
Notably, this doesn't handle things like failing if the
decrease would cause a requirement to no longer be fulfilled.
This is a complex problem, and will get more complex once I
add the grants system, but the general idea is as follows:

1. Before making any changes, make a copy of the model.
2. Apply the changes, propagate any consequences
3. Do a full validation of the new model.
4. If this fails, restore the previous model.

This _sounds_ overly complex, particularly since the only
obviously consistent way to implement `can_decrease` is to do all
of the above, throw away the result model, and report whether
it succeeded or failed. However, that's exactly how the Alliance CMA
is implemented (more or less), so in theory it should work here
as well. The Alliance ruleset doesn't really have a concept of
propagating consequences (everything's a skill, so there's no need
for skills that grant other skills) so it's less complex in that
regard, so we could still encounter complexity issues here.