Rename "classfeature" to "subfeature".
The concept can potentially be reused for other cases where
a feature grants a choice between options or some extra benefit
that occurs after a later set of requirements are met.

Example: The Extended Capacity skills add a spell slot of a sphere
of your choice each time you take them. One way to represent this
could be that each rank of Extended Capacity grants a choice that can
be filled by a "Bonus Arcane Slot" or "Bonus Divine Slot" subfeature.
(Alternatively, each could be represented as a pair of skills that
has a shared max ranks pool, to simulate it being a single skill,
or an Option skill that similarly has max ranks across options).