SP Tracking, part 1
This is the easy part: Put an `sp` field on awards, and tally it on player records.

I have a mild concern that SP transactions will clog up PlayerRecord.awards. In the current
implementation, ALL awards go in here, but but we can potentially optimize this. Of course,
this is only a hypothetical concern, so I don't need to solve it just now.